chore: move BuildReport to Editor module
