--- Internal/ErrorReporter/Editor/ErrorLogExtensions.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 82f79b46084a4d019b8993ff726012e6
-timeCreated: 1682930245
